Put styles into Address Modal, remove separate style file
